integration: split out wrapper python and add some helpers
The helpers get-config.py and get-structure.py collect and pretty
print the config and structure built-into the viaems binary.
