+++ integration/get-config.py
@@ -1,0 +1,10 @@
+from viaems import ViaemsWrapper
+import json
+
+conn = ViaemsWrapper("obj/hosted/viaems")
+conn.start()
+
+structure = conn.get(path=[])
+print(json.dumps(structure, indent=2))
+
+conn.kill()
